refactor: remove old Chain impl
`Chain` is an exploration into type-based commit behavior. Where in
an inner value is wrapped, with a hierarchy of types which commit
a tree of changes.

This model was difficult and assumed control of the commit process
from interfaces (Map, Bytes, etc).

Functionality of tree-updates have been primarily moved to
`crate::Path` and `crate::path::Segment`.