feat: impl Caching API and Usage
To facilitate shared byte array access with a cache, a `CacheRead`
and `CacheWrite` API has been added. The goal is to allow a
low/zero cost deserialzer ala Rkyv to deserialize from a cached
byte array.

Ultimately, `Arc<Vec<u8>>` is expected to be the storage type,
as the design required interior mutability.